+++ ncflag/flag_wrapper.py
@@ -1,5 +1,4 @@
 import numpy as np
-from six import string_types
 
 
 class FlagWrap(object):
@@ -15,8 +14,8 @@ class FlagWrap(object):
 
         :type flags: np.ndarray
         :param flags: array, flag values indicated
-        :type flag_meanings: list[string_types] | string_types
-        :param flag_meanings: list[string_types] | string_types, flag_meaning, string name of each designated flag_meaning
+        :type flag_meanings: list[str] | str
+        :param flag_meanings: list[str] | str, flag_meaning, string name of each designated flag_meaning
         :type flag_values: np.array
         :param flag_values: array, value of flag indicating corresponding flag_meaning set
         :type flag_masks: np.array
@@ -32,7 +31,7 @@ class FlagWrap(object):
         else:
             self.flags = np.array(flags)
 
-        if isinstance(flag_meanings, string_types):
+        if isinstance(flag_meanings, str):
             self.flag_meanings = flag_meanings.split()  # split on spaces
         else:
             assert isinstance(
@@ -132,7 +131,7 @@ class FlagWrap(object):
         """
         Get an array of booleans, same length as flags, at each index indicating if flag_meaning was set.
 
-        :type flag_meaning: string_types | list[string_types] | tuple[string_types]
+        :type flag_meaning: str | list[str] | tuple[str]
         :param flag_meaning: flag meaning(s) to be looked up
         :rtype: np.array
         :return: array of booleans where flag_meaning(s) is(are) set
@@ -199,7 +198,7 @@ class FlagWrap(object):
         """
         Returns True or False if flag_meaning set at index i?
         
-        :type flag_meaning: string_types
+        :type flag_meaning: str
         :param flag_meaning: flag meaning intended to be set
         :type i: int
         :param i: index in wrapped flags array to inspect.
@@ -221,7 +220,7 @@ class FlagWrap(object):
         :param i: the index to examine
         :type exit_on_good: bool
         :param exit_on_good: shortcut, return good as soon as good is found
-        :rtype: list[string_types]
+        :rtype: list[str]
         :return: a list of flags_meanings set at index i
         """
         flags_set = []
@@ -249,7 +248,7 @@ class FlagWrap(object):
         History: dealing with misspelled flag_meanings that will eventually be fixed, but now we must think
         about making the code work properly, transparently over time, over the point when the input is fixed.
 
-        :type options: list[string_types]
+        :type options: list[str]
         :param options: list of potential flag_meanings to seek
         :rtype: np.array
         :return: array of booleans indicating where first flag_meaning found is set.
@@ -271,7 +270,7 @@ class FlagWrap(object):
         flags where the whole value indicates the flag_meaning, so zeroing the target bits before setting
         on should_be_set will clear any previous flag_meanings or fill values that were set.
 
-        :type flag_meaning: string_types
+        :type flag_meaning: str
         :param flag_meaning: flag meaning intended to be set
         :type should_be_set: np.array
         :param should_be_set: array of booleans to set
@@ -302,7 +301,7 @@ class FlagWrap(object):
         of the flag_meaning within the bit vec while leaving the rest set or unset as they were.
         Then, OR the flag value onto the target, preserves all other independent flags set.
         
-        :type flag_meaning: string_types
+        :type flag_meaning: str
         :param flag_meaning: flag meaning intended to be set
         :type i: int
         :param i: index at which to set flag_meaning
@@ -329,7 +328,7 @@ class FlagWrap(object):
         
         Note: Raises ValueError if flag_meaning is not found.
 
-        :type flag_meaning: string_types
+        :type flag_meaning: str
         :param flag_meaning: string flag name to return value of
         :rtype: int
         :return: value of flag that sets flag_meaning
@@ -349,7 +348,7 @@ class FlagWrap(object):
 
         Note: Raises ValueError if flag_meaning not found.
         
-        :type flag_meaning: string_types
+        :type flag_meaning: str
         :param flag_meaning: string flag name to return corrsponding mask of
         :rtype: int
         :return: flag_mask value corresponding to flag_meaning
@@ -364,7 +363,7 @@ class FlagWrap(object):
         Convenience function to handle common use case of checking if a
         flag_meaning is valid for the wrapped flag.
 
-        :type flag_meaning: string_types
+        :type flag_meaning: str
         :param flag_meaning: string flag name to test existence of.
         :rtype: bool
         :return: if flag_meaning is a valid meaning for the flag.