+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  - Change reduce method to exclude _entire_ flag vectors (all bits) from reduction if any bits in
    the exclude_mask are set.
  - Make FlagWrap.flag_meanings "public" (prev. FlagWrap._flag_meanings)
+ - Adds name attribute, available from init as kwarg, for misc use.
 
 # v0.2.6 - 2020 04 03
  - Bug fix: cli: fix `--show_flags` option not working for Python 3