+++ CHANGELOG.md
@@ -1,10 +1,10 @@
-# unreleased
- - New method: `FlagWrap.is_valid_meaning` to test if a meaning is valid without accessing
-    the internal _flag_meanings attribute.
+# v0.3.0 - 2020 12 11
+ - New method: `FlagWrap.is_valid_meaning` for convenience.
  - Change reduce method to exclude _entire_ flag vectors (all bits) from reduction if any bits in
    the exclude_mask are set.
  - Make FlagWrap.flag_meanings "public" (prev. FlagWrap._flag_meanings)
  - Adds name attribute, available from init as kwarg, for misc use.
+ - Drops support for Python 2.7
 
 # v0.2.6 - 2020 04 03
  - Bug fix: cli: fix `--show_flags` option not working for Python 3